админка с разделами

+++ database/migrations/2023_04_18_100234_create_posts_table.php
@@ -17,8 +17,8 @@ class CreatePostsTable extends Migration
             $table->id();
             $table->string('title');
             $table->text('text');
-            $table->bigIncrements('author_id');
-            $table->bigIncrements('theme_id');
+            $table->integer('author_id');
+            $table->integer('theme_id');
             $table->timestamps();
         });
     }
